# swagger2.yaml
swagger: '2.0'
info:
  title: Adminiculum API
  description: Legal Workflow Platform API
  version: '1.0'
host: adminiculumaustriaeast-01.azurewebsites.net
basePath: /
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
  /api/v1/auth/login:
    post:
      summary: Login
      description: Bejelentkezés
      operationId: Login
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
                format: email
              password:
                type: string
      responses:
        '200':
          description: Login successful
          schema:
            type: object
            properties:
              accessToken:
                type: string
              refreshToken:
                type: string
              expiresIn:
                type: integer
              user:
                type: object
                properties:
                  id:
                    type: string
                  email:
                    type: string
                  name:
                    type: string
                  role:
                    type: string
  /api/v1/auth/register:
    post:
      summary: Register
      description: Regisztráció (teszteléshez)
      operationId: Register
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            required:
              - email
              - password
              - name
              - role
            properties:
              email:
                type: string
              password:
                type: string
              name:
                type: string
              role:
                type: string
                enum:
                  - ADMIN
                  - PARTNER
                  - LAWYER
                  - TRAINEE
                  - LEGAL_ASSISTANT
      responses:
        '201':
          description: Registration successful
  /api/v1/cases:
    get:
      summary: GetCases
      description: Ügyek listázása
      operationId: GetCases
      produces:
        - application/json
      responses:
        '200':
          description: List of cases
    post:
      summary: CreateCase
      description: Új ügy létrehozása
      operationId: CreateCase
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
      responses:
        '201':
          description: Case created
...
  /api/v1/cases/{id}:
    get:
      summary: GetCaseById
      description: Ügy részletei
      operationId: GetCaseById
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Case details
  /api/v1/cases/{id}/status:
    patch:
      summary: UpdateCaseStatus
      description: Ügy státusz frissítése
      operationId: UpdateCaseStatus
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          required: true
          type: string
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              status:
                type: string
      responses:
        '200':
          description: Status updated
...
  /api/v1/contracts/case/{caseId}:
    get:
      summary: GetCaseContracts
      description: Ügy szerződései
      operationId: GetCaseContracts
      produces:
        - application/json
      parameters:
        - name: caseId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: Case contracts
...
securityDefinitions:
  oauth2-auth:
    type: oauth2
    flow: accessCode
    tokenUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/token
    authorizationUrl: https://login.microsoftonline.com/18b56834-dfea-4931-bdf8-e5ebb0cb4e0f/oauth2/v2.0/authorize
    scopes:
      api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user: Access as user
security:
  - oauth2-auth:
      - api://a1e8b8a0-7690-4d09-9974-e4742d3de4e9/access_as_user